Give the induced weight an object, and its own file
`opening.rs` held three concerns in 786 lines: committing a level, the
Ligerito prover, and four free functions building the induced weight.

Those four all took `level` first, and all four carried the same inputs
around: the level, the transform, the query indices and the batching
coefficient. Inputs that always travel together are state, not
parameters. So they become `InducedWeight`, in its own file, with the
two builds and the selection rule as methods:

    InducedWeight::new(level, ntt, indices, alpha).build(alloc)

The rule reads `indices.len()` rather than `level.n_queries` now. They
are equal by construction, but the row count is what the cost actually
depends on, so the method reads the thing it means.

`by_rows` and `by_adjoint` stay reachable so the equality tests can
compare the two derivations directly, and they move to sit beside the
code they test. `transpose` became `adjoint` throughout: the route is
the encoder's adjoint, of which the transposed transform is one of three
steps, and the old name invited confusing it with the inverse.
