+++ crates/iop-prover/src/ligerito/mod.rs
@@ -8,6 +8,7 @@
 
 pub mod channel;
 pub mod compiler;
+mod induced_weight;
 mod opening;
 
 pub use opening::LigeritoProver;
