+++ crates/iop-prover/src/ligerito/opening.rs
@@ -13,7 +13,7 @@ use std::iter::zip;
 use binius_compute::{Allocator, GlobalAllocator};
 use binius_core::word::Word;
 use binius_field::{BinaryField, PackedField};
-use binius_iop::ligerito::{InducedBasis, LigeritoLevel, LigeritoParams};
+use binius_iop::ligerito::{LigeritoLevel, LigeritoParams};
 use binius_ip::mlecheck;
 use binius_ip_prover::sumcheck::{
 	bivariate_product_evaluator::bivariate_product_prover,
@@ -24,10 +24,11 @@ use binius_math::{
 	FieldBuffer, FieldSlice, FieldVec,
 	inner_product::inner_product_packed,
 	multilinear::{MultilinearMut, hypercube::Hypercube},
-	ntt::{AdditiveNTT, domain_context::GaoMateerOnTheFly},
+	ntt::AdditiveNTT,
 	reed_solomon::ReedSolomonCode,
 };
 
+use super::induced_weight::InducedWeight;
 use crate::{
 	fri::{BrakedownOracleProver, ProxTestOracleProver},
 	merkle_channel::MerkleIPProverChannel,
@@ -81,129 +82,6 @@ where
 	let commitment = channel.send_merkle_commitment(codeword.as_view(), 1 << level.log_lanes);
 
 	BrakedownOracleProver::new(codeword, commitment, 0)
-}
-
-/// The weight vector `level`'s opened rows induce, built over that level's codeword domain.
-///
-/// The prover needs every entry of it.
-/// A verifier reaches the same weight through the induced basis's closed form instead.
-/// So this is the one place the ladder holds a query position as a number rather than a word.
-/// It is also the only reason the proving loop pins the channel's word type.
-///
-/// Two builds reach the same vector, and which is cheaper turns on the level's shape.
-///
-/// ## Preconditions
-///
-/// * `ntt`'s domain covers the level's codeword domain.
-fn induced_weight<F, P, NTT, A>(
-	level: &LigeritoLevel,
-	ntt: &NTT,
-	indices: &[Word],
-	alpha: F,
-	alloc: &A,
-) -> FieldVec<P, A>
-where
-	F: BinaryField,
-	P: PackedField<Scalar = F>,
-	NTT: AdditiveNTT<Field = F> + Sync,
-	A: Allocator,
-{
-	let indices = indices
-		.iter()
-		.map(|index| index.as_u64() as usize)
-		.collect::<Vec<_>>();
-
-	if transposed_build_is_cheaper(level) {
-		induced_weight_by_transpose(level, ntt, &indices, alpha, alloc)
-	} else {
-		induced_weight_by_rows(level, &indices, alpha, alloc)
-	}
-}
-
-/// Builds the induced weight by expanding one opened row at a time.
-///
-/// A row is a tensor of `log_msg_cols` factors, so expanding it costs `2^log_msg_cols` products.
-/// Accumulating it into the running vector costs as many multiplications again.
-/// For `t` opened rows the whole build is therefore `2 * t * 2^log_msg_cols` multiplications.
-///
-/// This is the reference the transposed build is tested against.
-fn induced_weight_by_rows<F, P, A>(
-	level: &LigeritoLevel,
-	indices: &[usize],
-	alpha: F,
-	alloc: &A,
-) -> FieldVec<P, A>
-where
-	F: BinaryField,
-	P: PackedField<Scalar = F>,
-	A: Allocator,
-{
-	let domain_context = GaoMateerOnTheFly::generate(level.log_codeword_len());
-	let dense = InducedBasis::new(&domain_context, level.log_msg_cols, indices, alpha).to_dense();
-	FieldBuffer::from_values_in(alloc, &dense)
-}
-
-/// Builds the induced weight as one pass of the encoder's adjoint.
-///
-/// The weight is defined by what it does to a message:
-///
-/// ```text
-///     <w, m> = sum_i alpha^i * encode(m)[q_i] = <a, encode(m)>,   a[q_i] = alpha^i
-/// ```
-///
-/// The right-hand side is the encoder's adjoint identity read backwards.
-/// So `w` is that adjoint applied to the sparse weight `a`.
-///
-/// The cost is `log_msg_cols` butterfly layers over `2^(log_msg_cols + log_inv_rate)` entries.
-/// Nothing in it grows with the number of rows opened.
-///
-/// ## Preconditions
-///
-/// * `ntt`'s domain covers the level's codeword domain.
-fn induced_weight_by_transpose<F, P, NTT, A>(
-	level: &LigeritoLevel,
-	ntt: &NTT,
-	indices: &[usize],
-	alpha: F,
-	alloc: &A,
-) -> FieldVec<P, A>
-where
-	F: BinaryField,
-	P: PackedField<Scalar = F>,
-	NTT: AdditiveNTT<Field = F> + Sync,
-	A: Allocator,
-{
-	// A position can be drawn twice, so the powers landing on it accumulate rather than overwrite.
-	let mut weights = FieldBuffer::zeros_in(alloc, level.log_codeword_len());
-	for (&index, power) in zip(indices, alpha.powers()) {
-		weights.set(index, weights.get(index) + power);
-	}
-
-	let code = ReedSolomonCode::<F>::new(level.log_msg_cols, level.log_inv_rate);
-	code.encode_batch_transpose(ntt, weights.as_mut_view(), 0, alloc)
-}
-
-/// Whether the transposed build beats the row-by-row one at this level's shape.
-///
-/// Both builds produce `2^cols` entries, so dividing their costs by `2^cols` leaves
-///
-/// ```text
-///     row by row   t row entries
-///     transposed   cols * 2^rate layer entries
-/// ```
-///
-/// where `t` is the number of opened rows.
-/// Neither cost depends on the level's lane count, which is why that field is not read here.
-///
-/// A row entry and a layer entry are not the same price, so the comparison needs a constant.
-/// Timing both builds over 128-bit binary field elements puts a row entry at 2.9 ns.
-/// It puts a layer entry at 1.6 ns, steady to a few percent from `2^9` up to `2^24` entries.
-/// The ratio of the two, 1.9, is what the integers below encode.
-const fn transposed_build_is_cheaper(level: &LigeritoLevel) -> bool {
-	// The two costs above, each scaled by 10 so the measured 1.9 is an integer ratio.
-	let row_cost = 19 * level.n_queries;
-	let transposed_cost = 10 * (level.log_msg_cols << level.log_inv_rate);
-	transposed_cost < row_cost
 }
 
 /// Proves a Ligerito opening against a committed ladder of Reed-Solomon codewords.
@@ -365,7 +243,7 @@ where
 
 			if next.is_some() {
 				let beta: F = channel.sample();
-				let mut glued = induced_weight(level, self.ntt, &indices, alpha, alloc);
+				let mut glued = InducedWeight::new(level, self.ntt, &indices, alpha).build(alloc);
 
 				// The claim the opened rows make is the induced weight paired with the message
 				// they folded to, which is what the verifier reads off as `enforced_sum`.
@@ -410,8 +288,7 @@ mod tests {
 		test_utils::random_field_buffer,
 	};
 	use binius_transcript::{ProverTranscript, VerifierTranscript, fiat_shamir::HasherChallenger};
-	use proptest::prelude::*;
-	use rand::{RngExt, SeedableRng, rngs::StdRng};
+	use rand::{SeedableRng, rngs::StdRng};
 
 	use super::*;
 	use crate::merkle_channel::ProverMerkleTranscriptChannel;
@@ -639,144 +516,6 @@ mod tests {
 
 				assert_eq!(params.proof_size(&scheme), written, "{lanes:?} n_queries={n_queries}");
 			}
-		}
-	}
-
-	/// Both builds of the induced weight, over one level shape and one set of query indices.
-	fn both_builds(
-		log_msg_cols: usize,
-		log_inv_rate: usize,
-		indices: &[usize],
-		alpha: B128,
-	) -> (FieldBuffer<B128>, FieldBuffer<B128>) {
-		// The lane count does not enter either build, so any value serves here.
-		let level = LigeritoLevel {
-			log_msg_cols,
-			log_lanes: 1,
-			log_inv_rate,
-			n_queries: indices.len(),
-		};
-		let ntt = NeighborsLastSingleThread::new(GaoMateerOnTheFly::<B128>::generate(
-			level.log_codeword_len(),
-		));
-		(
-			induced_weight_by_rows(&level, indices, alpha, &GlobalAllocator),
-			induced_weight_by_transpose(&level, &ntt, indices, alpha, &GlobalAllocator),
-		)
-	}
-
-	/// The two builds must agree entry for entry, not merely at one evaluation point.
-	///
-	/// The transposed build reaches the weight through the encoder's adjoint.
-	/// The row build reaches it through the tensor expansion of each generator row.
-	/// Nothing but a test connects the two derivations.
-	#[test]
-	fn the_transposed_build_matches_the_row_build_at_the_boundaries() {
-		let alpha = B128::new(0x9e3779b97f4a7c15);
-
-		// A message of one column, which leaves the transform no layer to run at all.
-		let (rows, transposed) = both_builds(0, 3, &[0, 5, 5], alpha);
-		assert_eq!(rows, transposed);
-
-		// No rows opened, which induces the all-zero weight rather than an empty one.
-		let (rows, transposed) = both_builds(4, 2, &[], alpha);
-		assert_eq!(rows, transposed, "no rows");
-		assert_eq!(rows, FieldBuffer::zeros(4));
-
-		// Every rate the ladder search may pick, at one row and at a row drawn three times.
-		// Sampling is with replacement.
-		// So a repeated index must add its powers rather than overwrite them.
-		for log_inv_rate in 1..=8 {
-			let (rows, transposed) = both_builds(3, log_inv_rate, &[6], alpha);
-			assert_eq!(rows, transposed, "one row at rate {log_inv_rate}");
-
-			let (rows, transposed) = both_builds(3, log_inv_rate, &[2, 2, 2], alpha);
-			assert_eq!(rows, transposed, "a repeated row at rate {log_inv_rate}");
-		}
-	}
-
-	/// The selection rule, at every level shape the ladder search picks at 96-bit security.
-	///
-	/// The last column is the row build's measured time over the transposed build's.
-	/// A value above one is therefore a shape where the transposed build won.
-	/// The rule must agree with that comparison at every shape.
-	#[test]
-	fn the_selection_rule_follows_the_measured_crossover() {
-		// (log_msg_cols, log_inv_rate, n_queries, measured speedup of the transposed build)
-		let measured: &[(usize, usize, usize, f64)] = &[
-			(9, 1, 232, 26.43),
-			(9, 3, 116, 3.86),
-			(9, 5, 101, 0.88),
-			(9, 6, 99, 0.41),
-			(10, 3, 116, 3.06),
-			(10, 4, 106, 1.39),
-			(10, 5, 101, 0.66),
-			(10, 6, 99, 0.32),
-			(10, 8, 97, 0.08),
-			(11, 4, 106, 1.17),
-			(11, 6, 99, 0.26),
-			(12, 4, 106, 1.12),
-			(12, 5, 101, 0.52),
-			(13, 2, 142, 5.15),
-			(13, 4, 106, 0.94),
-			(13, 5, 101, 0.43),
-			(13, 6, 99, 0.20),
-			(14, 1, 232, 15.21),
-			(14, 7, 98, 0.09),
-			(15, 4, 106, 0.78),
-			(15, 5, 101, 0.37),
-			(16, 2, 142, 4.12),
-			(16, 4, 106, 0.74),
-			(17, 1, 232, 11.75),
-			(17, 4, 106, 0.74),
-			(17, 5, 101, 0.34),
-			(18, 6, 99, 0.16),
-			(19, 4, 106, 0.75),
-			(19, 5, 101, 0.34),
-			(20, 2, 142, 3.46),
-			(21, 1, 232, 10.95),
-			(21, 4, 106, 0.62),
-			(22, 1, 232, 10.36),
-			(23, 3, 116, 1.26),
-			(24, 1, 232, 9.60),
-			(24, 2, 142, 2.96),
-		];
-		for &(log_msg_cols, log_inv_rate, n_queries, speedup) in measured {
-			let level = LigeritoLevel {
-				log_msg_cols,
-				log_lanes: 1,
-				log_inv_rate,
-				n_queries,
-			};
-			assert_eq!(
-				transposed_build_is_cheaper(&level),
-				speedup > 1.0,
-				"cols={log_msg_cols} rate={log_inv_rate} t={n_queries} speedup={speedup}"
-			);
-		}
-	}
-
-	proptest! {
-		/// The two builds must agree over the whole shape space, not the shapes a fixed test picks.
-		///
-		/// Indices are drawn with replacement, exactly as the channel draws them.
-		/// A run can therefore land on the same row twice.
-		#[test]
-		fn the_transposed_build_matches_the_row_build(
-			seed: u64,
-			log_msg_cols in 0usize..7,
-			log_inv_rate in 1usize..5,
-			n_queries in 0usize..24,
-		) {
-			let mut rng = StdRng::seed_from_u64(seed);
-			let log_codeword_len = log_msg_cols + log_inv_rate;
-			let indices = (0..n_queries)
-				.map(|_| rng.random_range(0..1usize << log_codeword_len))
-				.collect::<Vec<_>>();
-			let alpha = B128::random(&mut rng);
-
-			let (rows, transposed) = both_builds(log_msg_cols, log_inv_rate, &indices, alpha);
-			prop_assert_eq!(rows, transposed);
 		}
 	}
 }
